+++ src/pynwb/ophys.py
@@ -1,4 +1,3 @@
-from collections.abc import Iterable
 import numpy as np
 import warnings
 
@@ -252,7 +251,7 @@ class TwoPhotonSeries(ImageSeries):
     @docval(*get_docval(ImageSeries.__init__, 'name'),  # required
             {'name': 'imaging_plane', 'type': ImagingPlane, 'doc': 'Imaging plane class/pointer.'},  # required
             *get_docval(ImageSeries.__init__, 'data', 'unit', 'format'),
-            {'name': 'field_of_view', 'type': (Iterable, TimeSeries), 'shape': ((2, ), (3, )),
+            {'name': 'field_of_view', 'type': ('array_data', 'data', TimeSeries), 'shape': ((2, ), (3, )),
              'doc': 'Width, height and depth of image, or imaged area (meters).', 'default': None},
             {'name': 'pmt_gain', 'type': float, 'doc': 'Photomultiplier gain.', 'default': None},
             {'name': 'scan_line_rate', 'type': float,